+++ docs/CHANGELOG.md
@@ -1,16 +1,20 @@
 # Changelog
 
+### 2.2.26 changelog
+
+- `Improvements` *Caret* — Improvements of the caret behaviour: arrows, backspace and enter keys better handling.
+
 ### 2.2.25 changelog
 
-- `New` *Autofocus — Now you can set focus at Editor after page has been loaded
+- `New` *Autofocus* — Now you can set focus at Editor after page has been loaded
 
 ### 2.2.24 changelog
 
-- `Improvements` *Paste handling — minor paste handling improvements
+- `Improvements` *Paste* handling — minor paste handling improvements
 
 ### 2.2.23 changelog
 
-- `New` *Shortcuts — copy and cut Blocks selected by CMD+A
+- `New` *Shortcuts* — copy and cut Blocks selected by CMD+A
 
 ### 2.2—2.7 changelog